+++ sdk/migration/azure-resourcemanager-migration-assessment/src/main/java/com/azure/resourcemanager/migration/assessment/models/AzureAvsNodeType.java
@@ -22,6 +22,21 @@ public final class AzureAvsNodeType extends ExpandableStringEnum<AzureAvsNodeTyp
     public static final AzureAvsNodeType AV36 = fromString("AV36");
 
     /**
+     * Static value AV36P for AzureAvsNodeType.
+     */
+    public static final AzureAvsNodeType AV36P = fromString("AV36P");
+
+    /**
+     * Static value AV52 for AzureAvsNodeType.
+     */
+    public static final AzureAvsNodeType AV52 = fromString("AV52");
+
+    /**
+     * Static value AV64 for AzureAvsNodeType.
+     */
+    public static final AzureAvsNodeType AV64 = fromString("AV64");
+
+    /**
      * Creates a new instance of AzureAvsNodeType value.
      * 
      * @deprecated Use the {@link #fromString(String)} factory method.