+++ sdk/migration/azure-resourcemanager-migration-assessment/src/main/java/com/azure/resourcemanager/migration/assessment/models/Percentile.java
@@ -8,7 +8,7 @@ import com.azure.core.util.ExpandableStringEnum;
 import java.util.Collection;
 
 /**
- * Defines values for Percentile.
+ * Percentile for Performance Data.
  */
 public final class Percentile extends ExpandableStringEnum<Percentile> {
     /**