+++ sdk/migration/azure-resourcemanager-migration-assessment/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## 1.0.0-beta.1 (2025-02-26)
 
+- Azure Resource Manager Migration Assessment client library for Java. This package contains Microsoft Azure SDK for Migration Assessment Management SDK. Azure Migrate Resource Provider management API. Package tag package-preview-2024-01. For documentation on how to use this package, please see [Azure Management Libraries for Java](https://aka.ms/azsdk/java/mgmt).
 - Azure Resource Manager Migration Assessment client library for Java. This package contains Microsoft Azure SDK for Migration Assessment Management SDK. Azure Migrate Resource Provider management API. Package tag package-2023-03. For documentation on how to use this package, please see [Azure Management Libraries for Java](https://aka.ms/azsdk/java/mgmt).
 ### Features Added
 
 - Initial release for the azure-resourcemanager-migration-assessment Java SDK.
+